+++ packages/xchain-doge/__mocks__/sochain.ts
@@ -1,4 +1,4 @@
-import axios from 'axios'
+const axios = require('axios')
 import MockAdapter from 'axios-mock-adapter'
 
 const mock = new MockAdapter(axios)